Fix java.security.cert.CertPathValidatorException: Trust anchor for certification path not found.
Exception occurs on Android devices that fail to properly parse the
DER-encoded trusted-certs.crt file if it contains comments.

Note that this fix was actually applied in
d9621257a14957f88080643272a74f0c47b582f7, but we document the error
case better here (and update the Android example so it will fail next
time this happens).