Moving settings.gradle to standalone-settings.gradle.
Android Studio will not detect submodules that contain a settings.gradle file. This prevents Android
Studio from finding this project when it is configured as a subproject of dbapp-android. Note that
building dbapp-android from the commandline works fine when settings.gradle exists. It is only when
using Android Studio that the build fails.

Until Android Studio is fixed (not likely), avoid adding settings.gradle file to this project.