Expose an additional jar that doesn't include OSGi manifest information
Currently doing a noop build of the SDK jar takes quite a while (for me, the
jar task itself takes about 2.5 seconds to determine that it's up to date). I
dug into this and it turns out that this is due to an inefficiency in the OSGi
plugin when determining if the OSGi manifest is up to date.

On my machine, a noop jarWithoutOsgi task is ~50x faster, taking about 0.05
seconds.