Allow generation of observable routes via Rx
Summary:

Adds a new flag (--observable-routes) to java.stoneg.py that causes
it to generate (currently only for rpc routes) methods that return Singles
and Completables allowing for interop with Rx.

It does this by wrapping calls to the non-rx methods so the total
amount of change in the generate isn't too large.

Test Plan:

Test with an example version of account-info. Integrate with the Dropbox
Android app and test an end to end flow. The logic is guarded by a flag
and so users that don't opt in should not be affected.