Fix for DbxRefreshResult Expires In
This change replaces in all the DbxClient implementors the `refreshAccessToken` function to make sure it is created with the proper Expires In param.
DbxRefreshResult expects Expires In Seconds, which is the time remaining for the token until expired.
However, we were passing in Expires At, which is the point in time where the token will be expired, in Millis.

The main idea is ExpiresIn seconds = (ExpiresAt millis - current time millis) / 1000

Modified the test cases to validate this logic and added an Offset because there is no way to inject the issued time.